SAR-7764: Updated Annual Costs Limited view and added view spec
